requirements.txt
fastapi
pydantic
uvicorn
pytesseract
gunicorn
Pillow
python-multipart